use `CommonModel<T>` for all models
