+++ core/src/models/Export.ts
@@ -1,19 +1,14 @@
 import {
-  Model,
   Table,
   Column,
-  CreatedAt,
-  UpdatedAt,
   AllowNull,
   Is,
   BelongsTo,
-  BeforeCreate,
   BeforeSave,
   ForeignKey,
   DataType,
   Default,
 } from "sequelize-typescript";
-import * as uuid from "uuid";
 import { Destination } from "./Destination";
 import { GrouparooRecord } from "./GrouparooRecord";
 import { plugin } from "../modules/plugin";
@@ -26,6 +21,7 @@ import { api, config } from "actionhero";
 import { ExportProcessor } from "./ExportProcessor";
 import { Errors } from "../modules/errors";
 import { PropertyTypes } from "./Property";
+import { CommonModel } from "../classes/commonModel";
 
 /**
  * The GrouparooRecord Properties in their normal data types (string, boolean, date, etc)
@@ -67,19 +63,10 @@ const STATE_TRANSITIONS = [
 ];
 
 @Table({ tableName: "exports", paranoid: false })
-export class Export extends Model {
+export class Export extends CommonModel<Export> {
   idPrefix() {
     return "exp";
   }
-
-  @Column({ primaryKey: true })
-  id: string;
-
-  @CreatedAt
-  createdAt: Date;
-
-  @UpdatedAt
-  updatedAt: Date;
 
   @AllowNull(false)
   @ForeignKey(() => Destination)
@@ -290,11 +277,6 @@ export class Export extends Model {
   @BeforeSave
   static async updateState(instance: Export) {
     await StateMachine.transition(instance, STATE_TRANSITIONS);
-  }
-
-  @BeforeCreate
-  static generateId(instance: Export) {
-    if (!instance.id) instance.id = `${instance.idPrefix()}_${uuid.v4()}`;
   }
 
   @BeforeSave
