+++ core/src/models/Session.ts
@@ -1,33 +1,15 @@
-import {
-  Model,
-  Table,
-  Column,
-  AllowNull,
-  ForeignKey,
-  CreatedAt,
-  UpdatedAt,
-  BeforeCreate,
-} from "sequelize-typescript";
-import * as uuid from "uuid";
+import { Table, Column, AllowNull, ForeignKey } from "sequelize-typescript";
 import { TeamMember } from "./TeamMember";
 import { APIData } from "../modules/apiData";
 import { Op } from "sequelize";
 import { api } from "actionhero";
+import { CommonModel } from "../classes/commonModel";
 
 @Table({ tableName: "sessions", paranoid: false })
-export class Session extends Model {
-  @Column({ primaryKey: true })
-  id: string;
-
+export class Session extends CommonModel<Session> {
   idPrefix() {
     return "ses";
   }
-
-  @CreatedAt
-  createdAt: Date;
-
-  @UpdatedAt
-  updatedAt: Date;
 
   @AllowNull(false)
   @Column
@@ -59,11 +41,6 @@ export class Session extends Model {
     const instance = await this.scope(null).findOne({ where: { id } });
     if (!instance) throw new Error(`cannot find ${this.name} ${id}`);
     return instance;
-  }
-
-  @BeforeCreate
-  static generateId(instance: Session) {
-    if (!instance.id) instance.id = `${instance.idPrefix()}_${uuid.v4()}`;
   }
 
   static async sweep() {
